+++ packages/react/AppShell/src/components/AppShellHeader/AppShellHeader.tsx
@@ -28,6 +28,4 @@ const AppShellHeader: React.FC<AppShellHeaderProps> = ({
   </Box>
 );
 
-AppShellHeader.displayName = "AppShell.Header";
-
 export { AppShellHeader };