+++ packages/react/FormField/src/FormField.tsx
@@ -46,5 +46,8 @@ FormField.Textarea = FormFieldTextarea;
 FormField.Input = FormFieldInput;
 
 FormField.displayName = "FormField";
+FormField.Input.displayName = "FormField.Input";
+FormField.Select.displayName = "FormField.Select";
+FormField.Textarea.displayName = "FormField.Textarea";
 
 export { FormField };