+++ packages/react/FormField/src/components/FormFieldInput/FormFieldInput.tsx
@@ -26,6 +26,4 @@ const FormFieldInput: React.FC<FormFieldInputProps> = ({
   </FormField>
 );
 
-FormFieldInput.displayName = "FormField.Input";
-
 export { FormFieldInput };