+++ packages/react/FormField/src/formField.types.ts
@@ -15,10 +15,16 @@ export interface FormFieldComponents {
 }
 
 export interface FormFieldProps extends Extend {
+  /** Optional label for the field component */
   label?: ReactNode;
+  /** Help text displaying optional hints or validation messages under the field */
   helpText?: string;
+  /** Icon supporting the help text message */
   helpIcon?: FC<IconProps>;
+  /** Appearance of the field and help text elements */
   appearance?: "danger" | "warning" | "success" | "none";
+  /** Control to conditionally show the help text and icon */
   showHelpText?: boolean;
+  /** Content of the field */
   children: ReactNode;
 }