+++ packages/react/Menu/src/components/MenuBody/menuBody.types.ts
@@ -1,5 +1,6 @@
 import { ReactNode, HTMLAttributes } from "react";
 
 export interface MenuBodyProps extends HTMLAttributes<HTMLElement> {
+  /** Content of the menu body */
   children: ReactNode;
 }