+++ packages/react/Page/src/components/PageBody/pageBody.types.ts
@@ -1,5 +1,6 @@
 import { ReactNode, HTMLAttributes } from "react";
 
 export interface PageBodyProps extends HTMLAttributes<HTMLElement> {
+  /** Content of the page body */
   children: ReactNode;
 }