+++ packages/react/Layout/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The Layout component is intended for building single or multi-column layouts using predefined typologies.
 
+## 2023-03-13 `1.1.1`
+
+- Add props documentation. ([#26](https://github.com/TiendaNube/nimbus-patterns/pull/26) by [@juanchigallego](https://github.com/juanchigallego))
+
 ## 2023-02-17 `1.1.0`
 
 ### 📚 3rd party library updates