+++ packages/react/ProductUpdates/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ProductUpdates component allows the user to showcase new features or updates in a non-distractive way by displaying the information in a floating popover with a high-contrast setting that catches the eye of the user.
 
+## 2023-03-13 `1.2.2`
+
+- Add props documentation. ([#26](https://github.com/TiendaNube/nimbus-patterns/pull/26) by [@juanchigallego](https://github.com/juanchigallego))
+
 ## 2023-03-02 `1.2.1`
 
 ### 📚 3rd party library updates